Added small comment to copy `c:\verdaccio\config.yaml` from `verdaccio\conf\default.yaml`
I spent quite a while failing to get verdaccio to works as a Windows service as a result of not actually knowing what I was doing regarding the config.yaml file. Fine if I was familiar with verdaccio but I was trying to work on the fly without reading the configuration documentation after having it work on the command line with 'npm install -g verdaccio`.